Moved DeathScreenDisplay, WinScreenDisplay and PauseMenuDisplay into a single package
