+++ source/core/src/main/com/csse3200/game/components/screens/WinScreenDisplay.java
@@ -1,4 +1,4 @@
-package com.csse3200.game.components.winscreenmenu;
+package com.csse3200.game.components.screens;
 
 import com.badlogic.gdx.graphics.g2d.SpriteBatch;
 import com.badlogic.gdx.scenes.scene2d.Actor;
